Dockerfile in the root reduces API callbacks.
Putting the Dockerfile in the root and using ENV allows us
to pass most configuration into the docker without a
number of insecure API calls.

It means that users will have to start with a checkout of
refstack instead of just the dockerfile, but that's expected.

I've also cleaned up the dashboard to unify the logged out and
logged in views, and made the "edit cloud" button work.

Change-Id: I25d249d9f95e56296593e2716001bd6921216ba1